verba_notes_backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.5
djangorestframework==3.15.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
